+++ src/util/bitcoincore_ext.rs
@@ -102,6 +102,10 @@ pub fn create_rpc_client(config: &crate::Config) -> Result<Client, crate::Error>
         builder = builder.auth(user, pass);
     }
 
+    if let Some(timeout) = config.bitcoind_timeout {
+        builder = builder.timeout(timeout);
+    }
+
     #[cfg(feature = "proxy")]
     if let Some(proxy_addr) = config.bitcoind_proxy {
         builder = builder.proxy(proxy_addr);